Personality quiz
Implemented a boolean user aspect of quizCompleted of which then tells an ejs script to render html with or without the quiz questions. Made the questions, though ugly. Started client side JS for them just haventt added the Post to the server and the middleware to add the info to the uer's profile in our database.